refactor init setup into record-data
